:memo: rename constants => types
